mostly done

+++ Frontent Practice/src/pages/History.jsx
@@ -1,13 +1,19 @@
 import React, { useEffect, useState } from "react";
 import { useDispatch, useSelector } from "react-redux";
-import { deleteFromHsitory, fetchHistory, setHistory } from "../store/history.slice.js";
+import {
+  deleteFromHsitory,
+  fetchHistory,
+  setHistory,
+} from "../store/history.slice.js";
 import { formatDistanceToNowStrict } from "date-fns";
 import { Link } from "react-router-dom";
 import { RxCross2 } from "react-icons/rx";
 import formatTime from "../utils/formatTime";
 import { deleteHistory } from "../apis/user.apis.js";
+import formatNumber from "../utils/formatNumber.js";
 
 const History = () => {
+  const [isDeleteLocked, setIsDeleteLocked] = useState(false);
   let { data, loading } = useSelector((store) => store.history);
   let dispatch = useDispatch();
   let [error, setError] = useState("");
@@ -18,18 +24,32 @@ const History = () => {
     }
   }, []);
 
-  let deleteVideo = async (id) => {
-    let prevHistory = [...data]
+  const deleteVideo = async (id) => {
+    if (isDeleteLocked) return;
+    setIsDeleteLocked(true);
+
+    // Unlock after 0.5s regardless of API completion
+    setTimeout(() => {
+      setIsDeleteLocked(false);
+    }, 500);
+
+    const prevHistory = [...data];
     try {
       dispatch(deleteFromHsitory(id));
       await deleteHistory(id);
     } catch (error) {
       setError(error?.message);
-      setHistory(prevHistory)
+      dispatch(setHistory(prevHistory));
     }
   };
+
   return (
     <div className="flex flex-col items-center py-5">
+      {data?.length > 0 && (
+        <div className="w-full flex justify-start md:w-[536px] lg:w-[760px] xl:w-[980px] mb-2">
+          <h2 className="text-3xl">History</h2>
+        </div>
+      )}
       {loading ? (
         <>
           {[...Array(4)].map((_, index) => (
@@ -37,9 +57,9 @@ const History = () => {
               key={index}
               className="md:flex py-2.5 gap-5 w-full cursor-pointer sm:w-full md:w-[536px] lg:w-[760px] xl:w-[980px] animate-pulse"
             >
-              <div className="relative w-full aspect-video overflow-hidden rounded-lg bg-gray-800 flex justify-center items-center">
+              <div className="relative w-full aspect-video overflow-hidden rounded-md bg-gray-800 flex justify-center items-center">
                 <div className="h-full w-full bg-gray-700"></div>
-                <div className="absolute bottom-2 right-2 bg-gray-600 bg-opacity-75 text-white text-xs px-1 py-0.5 rounded w-10 h-4"></div>
+                <div className="absolute bottom-2 right-2 bg-gray-600 bg-opacity-75   px-1 py-0.5 rounded w-10 h-4"></div>
               </div>
 
               <div className="w-full md:w-[50%] flex flex-col justify-start+">
@@ -51,13 +71,6 @@ const History = () => {
                 <div className="flex items-center gap-2 mb-2">
                   <div className="w-9 h-9 bg-gray-700 rounded-full"></div>
                   <div className="h-4 bg-gray-600 rounded w-24"></div>
-                </div>
-
-                <div className="hidden md:block">
-                  <div className="space-y-2">
-                    <div className="h-4 bg-gray-700 rounded w-full"></div>
-                    <div className="h-4 bg-gray-700 rounded w-[90%]"></div>
-                  </div>
                 </div>
               </div>
             </div>
@@ -82,22 +95,22 @@ const History = () => {
                         className="h-full object-contain"
                       />
 
-                      <div className="absolute bottom-2 right-2 bg-black bg-opacity-75 text-white text-xs px-1 py-0.5 rounded">
+                      <div className="absolute text-xs md:text-sm bottom-2 right-2 bg-black bg-opacity-75   px-1 py-0.5 rounded">
                         {formatTime(item?.video?.duration)}
                       </div>
                     </Link>
                   </div>
 
-                  <div className="w-full md:w-[50%]">
+                  <div className="md:block hidden md:w-[50%] ">
                     <Link
                       to={`/app/dashboard/single-video/${item?.video?.owner._id}/${item?.video?._id}`}
                     >
-                      <h3 className="text-lg my-2 font-semibold leading-tight break-words overflow-hidden text-ellipsis line-clamp-2">
+                      <h3 className=" my-2 text-md md:text-lg font-semibold leading-tight break-words overflow-hidden text-ellipsis line-clamp-2">
                         {item?.video?.title}
                       </h3>
                     </Link>
-                    <p className="text-gray-400 text-sm mb-2">
-                      {item?.video?.views} views •{" "}
+                    <p className="text-gray-400 text-sm md:text-md mb-2">
+                      {formatNumber(item?.video?.views)} views •{" "}
                       {formatDistanceToNowStrict(
                         new Date(item?.video?.createdAt),
                         {
@@ -105,7 +118,7 @@ const History = () => {
                         },
                       )}
                     </p>
-                    <div className="flex items-center justify-between gap-2 mb-2">
+                    <div className="flex items-center justify-between gap-2">
                       <div className="flex items-center gap-2">
                         <Link
                           to={`/app/dashboard/single-channel/${item?.video?.owner._id}`}
@@ -119,7 +132,7 @@ const History = () => {
                         </Link>
 
                         <Link
-                          className="text-gray-400"
+                          className="text-gray-400 text-sm md:text-md"
                           to={`/app/dashboard/single-channel/${item?.video?.owner._id}`}
                         >
                           {item?.video?.owner?.name}
@@ -134,21 +147,59 @@ const History = () => {
                         <RxCross2 />
                       </button>
                     </div>
-                    <div className="hidden md:block">
-                      <p className="my-2 leading-tight text-gray-300 break-words overflow-hidden text-ellipsis line-clamp-2">
-                        {item?.video?.description}
-                      </p>
+                  </div>
+
+                  <div className="w-full  md:hidden mt-3 flex flex-nowrap gap-2">
+                    <div className="flex flex-col flex-1 overflow-hidden w-full">
+                      <h3 className="text-white  font-semibold leading-tight break-words overflow-hidden text-ellipsis line-clamp-2">
+                        {item?.video?.title}
+                      </h3>
+                      <div className="text-sm text-gray-400 flex md:block items-center gap-1">
+                        <Link
+                          to={`/app/dashboard/single-channel/${item?.video?.owner._id}`}
+                        >
+                          <p className="text-sm text-gray-400 md:mt-1 whitespace-nowrap overflow-hidden text-ellipsis hover:text-gray-100">
+                            {item?.video?.owner?.name}
+                          </p>
+                        </Link>
+                        <p className="block md:hidden">•</p>
+                        <div className="flex items-center gap-1">
+                          <p className="text-sm text-gray-400 whitespace-nowrap overflow-hidden text-ellipsis">
+                            {formatNumber(item?.video?.views)} views •{" "}
+                            {formatDistanceToNowStrict(
+                              new Date(item?.video?.createdAt),
+                              {
+                                addSuffix: true,
+                              },
+                            )}
+                          </p>
+                        </div>
+                      </div>
+                    </div>
+                    <div>
+                      <button
+                        onClick={() => deleteVideo(item?._id)}
+                        disabled={isDeleteLocked}
+                        className="bg-gray-700 hover:bg-gray-600 p-3 rounded-4xl disabled:opacity-50"
+                      >
+                        <RxCross2 />
+                      </button>
                     </div>
                   </div>
                 </div>
               ))}
+              {data?.length != 0 && (
+                <p className="text-center pt-2 text-gray-400">
+                  No more history to show.
+                </p>
+              )}
             </>
           ) : (
             <div className="flex flex-col items-center justify-center h-[80vh]">
-              <h2 className="text-xl md:text-2xl font-semibold ">
+              <h2 className=" text-4xl font-bold text-gray-300 ">
                 No history to show
               </h2>
-              <p className="text-sm text-gray-500 mt-2">
+              <p className="text-md text-gray-500 mt-2">
                 Start watching videos to build your watch history.
               </p>
             </div>
